+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+### Added
+
+- `flipperzero::gpio::i2c`, providing a Rust interface to the external 3.3V I2C
+  bus over GPIO pins C0 and C1, as well as the internal (power) I2C bus.
+
+### Changed
+
 - Migrated to SDK 39.1 (firmware 0.92.2).
 
 ## [0.11.0]